requirements.txt
pandas >= 1.1.1
numpy >= 1.19.1
google-cloud-bigquery >= 1.27.2
google-cloud-bigquery-storage >= 1.0.0
google-cloud-storage >= 1.30.0
pandas-gbq >= 0.13.2